requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
frozenlist==1.8.0
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
googleapis-common-protos==1.74.0
httplib2==0.31.2
idna==3.11
multidict==6.7.1
propcache==0.4.1
proto-plus==1.27.2
protobuf==7.34.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.12.1
pyparsing==3.3.2
requests==2.33.1
twilio==9.10.4
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
yarl==1.23.0
